+++ lib/modules/cache/cache_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: cache
-// Updated at: 2026-04-17 11:18:07
-// Commit: chore(cache): update dependencies and cleanup unused imports in cache
+// Updated at: 2026-04-17 12:57:46
+// Commit: feat(cache): add bi-directional event bus handler for cache
 
 class cacheComponentState {
-  final String commitRef = '463';
-  final String lastMessage = 'chore(cache): update dependencies and cleanup unused imports in cache';
-  final DateTime timestamp = DateTime.parse('2026-04-17T11:18:07+05:00');
+  final String commitRef = '471';
+  final String lastMessage = 'feat(cache): add bi-directional event bus handler for cache';
+  final DateTime timestamp = DateTime.parse('2026-04-17T12:57:46+05:00');
 }